convert RlpEncoder into abstract class and add pre-sizing to RlpEncodable interface
