+++ ethers-core/src/main/kotlin/io/ethers/core/types/Address.kt
@@ -15,7 +15,6 @@ import io.ethers.rlp.RlpDecodable
 import io.ethers.rlp.RlpDecoder
 import io.ethers.rlp.RlpEncodable
 import io.ethers.rlp.RlpEncoder
-import io.ethers.rlp.RlpSizer
 import kotlin.random.Random
 
 /**
@@ -49,8 +48,6 @@ class Address(private val value: ByteArray) : RlpEncodable {
      * without copying.
      * */
     fun toByteArray() = value.copyOf()
-
-    override fun rlpEncodedSize() = RlpSizer.sizeOf(value)
 
     override fun rlpEncode(rlp: RlpEncoder) {
         rlp.encode(value)
@@ -96,7 +93,7 @@ class Address(private val value: ByteArray) : RlpEncodable {
         @JvmStatic
         fun computeCreate(sender: Address, nonce: Long): Address {
             // address.size + 8 (+ 1 RLP flag) = 29 bytes worst case
-            val rlp = RlpEncoder(sender.value.size + 9).encodeList {
+            val rlp = RlpEncoder.sized(sender.value.size + 9).encodeList {
                 encode(sender)
                 encode(nonce)
             }
