+++ ethers-core/src/main/kotlin/io/ethers/core/types/Hash.kt
@@ -14,7 +14,6 @@ import io.ethers.rlp.RlpDecodable
 import io.ethers.rlp.RlpDecoder
 import io.ethers.rlp.RlpEncodable
 import io.ethers.rlp.RlpEncoder
-import io.ethers.rlp.RlpSizer
 
 /**
  * 32-byte hash.
@@ -48,8 +47,6 @@ class Hash(private val value: ByteArray) : RlpEncodable {
      * without copying.
      * */
     fun toByteArray() = value.copyOf()
-
-    override fun rlpEncodedSize() = RlpSizer.sizeOf(value)
 
     override fun rlpEncode(rlp: RlpEncoder) {
         rlp.encode(value)
