+++ ethers-core/src/main/kotlin/io/ethers/core/types/transaction/TransactionSigned.kt
@@ -10,7 +10,6 @@ import io.ethers.rlp.RlpDecodable
 import io.ethers.rlp.RlpDecoder
 import io.ethers.rlp.RlpEncodable
 import io.ethers.rlp.RlpEncoder
-import io.ethers.rlp.RlpSizer
 import java.math.BigInteger
 
 /**
@@ -40,7 +39,7 @@ class TransactionSigned @JvmOverloads constructor(
     override val hash: Hash
         get() {
             if (_hash == null) {
-                val hashRlp = RlpEncoder()
+                val hashRlp = RlpEncoder.unsized()
                     .also { tx.rlpEncodeEnveloped(it, signature, true) }
                     .toByteArray()
 
@@ -109,30 +108,6 @@ class TransactionSigned @JvmOverloads constructor(
 
     override fun toString(): String {
         return "TransactionSigned(tx=$tx, signature=$signature)"
-    }
-
-    override fun rlpEncodedSize(): Int {
-        return with(RlpSizer) {
-            var size = 0
-            if (tx.type != TxType.Legacy) {
-                size += 1
-            }
-
-            size += sizeOfList {
-                var innerSize = 0
-                if (tx.type == TxType.Blob && (tx as TxBlob).sidecar != null) {
-                    innerSize += sizeOfList { tx.rlpFieldsSize() + signature.rlpEncodedSize() }
-                    innerSize += tx.sidecar!!.rlpEncodedSize()
-                } else {
-                    innerSize += tx.rlpFieldsSize()
-                    innerSize += signature.rlpEncodedSize()
-                }
-
-                innerSize
-            }
-
-            size
-        }
     }
 
     override fun rlpEncode(rlp: RlpEncoder) {
