+++ ethers-core/src/main/kotlin/io/ethers/core/types/transaction/TransactionUnsigned.kt
@@ -12,8 +12,6 @@ import java.math.BigInteger
  * An unsigned [Transaction] with functions for signing.
  * */
 sealed interface TransactionUnsigned : Transaction, RlpEncodable {
-    fun rlpFieldsSize(): Int
-
     /**
      * RLP encode transaction with optional signature according to [EIP-2718](https://eips.ethereum.org/EIPS/eip-2718)
      * enveloped format. If [hashEncoding] is true, the transaction is encoded for hash calculation, either signature
@@ -25,7 +23,7 @@ sealed interface TransactionUnsigned : Transaction, RlpEncodable {
      * Get hash used for signing the transaction.
      * */
     fun signatureHash(): ByteArray {
-        val encoder = RlpEncoder()
+        val encoder = RlpEncoder.unsized()
         rlpEncodeEnveloped(encoder, null, true)
         return Hashing.keccak256(encoder.toByteArray())
     }
