+++ ethers-core/src/main/kotlin/io/ethers/core/types/transaction/TxAccessList.kt
@@ -8,7 +8,6 @@ import io.ethers.core.types.Signature
 import io.ethers.rlp.RlpDecodable
 import io.ethers.rlp.RlpDecoder
 import io.ethers.rlp.RlpEncoder
-import io.ethers.rlp.RlpSizer
 import java.math.BigInteger
 
 /**
@@ -46,19 +45,6 @@ data class TxAccessList(
 
     override val blobVersionedHashes: List<Hash>?
         get() = null
-
-    override fun rlpFieldsSize(): Int {
-        return with(RlpSizer) {
-            sizeOf(chainId) +
-                sizeOf(nonce) +
-                sizeOf(gasPrice) +
-                sizeOf(gas) +
-                sizeOf(to) +
-                sizeOf(value) +
-                sizeOf(data) +
-                sizeOfList(accessList)
-        }
-    }
 
     override fun rlpEncodeEnveloped(rlp: RlpEncoder, signature: Signature?, hashEncoding: Boolean) {
         rlp.appendRaw(type.type.toByte())
