+++ ethers-rlp/src/main/kotlin/io/ethers/rlp/RlpEncodable.kt
@@ -5,11 +5,11 @@ interface RlpEncodable {
      * Return RLP encoded byte array of this class.
      */
     fun toRlp(): ByteArray {
-        val encoder = RlpEncoder(rlpEncodedSize())
-        return encoder.encode(this).toByteArray()
+        val sizer = RlpSizingEncoder().apply { rlpEncode(this) }
+        return RlpBufferEncoder(ByteArray(sizer.size()), isCorrectlySized = true)
+            .encode(this)
+            .toByteArray()
     }
-
-    fun rlpEncodedSize(): Int
 
     /**
      * Encode this class into provided [RlpEncoder].
